test : 다중 등록 `null` check test controller test 로 이동
